requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
pinecone-client
groq
pytubefix
streamlit